Refactor TodoDetailsPanel into TodoItemRUD component for better organization, enhancing readability and maintainability
